requirements.txt
Flask
Flask-Cors
numpy
opencv-python
scipy
Pillow
torch
torchaudio
torchvision
pytest
pytest-cov
requests